# Add benchmarks using the performance.json vmtests