--- a/jsontests/tests/arithmetic.rs
+++ b/jsontests/tests/arithmetic.rs
@@ -10,7 +10,7 @@ use jsontests::test_transaction;
 
 lazy_static! {
     static ref TESTS: Value =
-        serde_json::from_str(include_str!("files/vmArithmeticTest.json")).unwrap();
+        serde_json::from_str(include_str!("../res/files/vmArithmeticTest.json")).unwrap();
 }
 
 #[test] fn add0() { assert_eq!(test_transaction("add0", &TESTS["add0"], true), true); }