--- a/jsontests/tests/pushdupswap.rs
+++ b/jsontests/tests/pushdupswap.rs
@@ -10,7 +10,7 @@ use jsontests::test_transaction;
 
 lazy_static! {
     static ref TESTS: Value =
-        serde_json::from_str(include_str!("files/vmPushDupSwapTest.json")).unwrap();
+        serde_json::from_str(include_str!("../res/files/vmPushDupSwapTest.json")).unwrap();
 }
 
 #[test] fn dup1() { assert_eq!(test_transaction("dup1", &TESTS["dup1"], true), true); }