… jsontests/res/files/vmArithmeticTest.json
… jsontests/res/files/vmBitwiseLogicOperationTest.json
… jsontests/res/files/vmBlockInfoTest.json
… jsontests/res/files/vmEnvironmentalInfoTest.json
… jsontests/res/files/vmIOandFlowOperationsTest.json
… jsontests/res/files/vmLogTest.json
… jsontests/res/files/vmPerformanceTest.json
… jsontests/res/files/vmPushDupSwapTest.json
… jsontests/res/files/vmSha3Test.json
… jsontests/res/files/vmSystemOperationsTest.json