Also contains various fixes for `cargo bench`. It also add a new nix-shell file nightly.nix. Rust nightly is required to run benchmarks.

Current benchmarks:

```
test ackermann31       ... bench:   6,000,519 ns/iter (+/- 404,041)
test ackermann32       ... bench:  30,039,762 ns/iter (+/- 656,800)
test ackermann33       ... bench:  51,803,412 ns/iter (+/- 16,190,187)
test fibonacci10       ... bench:  10,333,038 ns/iter (+/- 1,547,309)
test fibonacci16       ... bench: 194,273,888 ns/iter (+/- 18,900,368)
test many_functions100 ... bench:  15,610,461 ns/iter (+/- 363,363)
```

For `fibonacci16` that's around 0.194 seconds for the whole operation. Compared with `evm`:

```
real	0m0.021s
user	0m0.019s
sys	0m0.001s
```